requirements.txt
ollama
mkdocs
python-docx
pymupdf
pillow
jinja2
